Change entity from int to long. Change ProjectBlueprint => Blueprint. Change repo from CrudRepo => JpaRepo (has both PagingRepo & CrudRepo).. Remove all paging repo and replace with equivalent in regular repo

+++ src/main/java/com/ntv/ntvcons_backend/entities/ExternalFile.java
@@ -15,7 +15,7 @@ public class ExternalFile extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "fileId", nullable = false)
-    private Integer fileId;
+    private Long fileId;
 
     @Column(name = "fileName", nullable = false, length = 100)
     private String fileName;
@@ -24,7 +24,7 @@ public class ExternalFile extends BaseEntity {
     private String fileLink;
 
     @Column(name = "fileTypeId", nullable = false)
-    private Integer fileTypeId;
+    private Long fileTypeId;
 
     @Column(name = "isDeleted", nullable = false)
     private Boolean isDeleted = false;
